[Bronze II] Title: 숫자, Time: 892 ms, Memory: 35952 KB, Score: 100 point -BaekjoonHub

+++ Java/백준/Bronze/10093. 숫자/숫자.java
@@ -4,21 +4,21 @@ public class Main
     public static void main(String args[])
     {
         Scanner scanner = new Scanner(System.in);
-        int a = scanner.nextInt();
-        int b = scanner.nextInt();
+        long a = scanner.nextLong();
+        long b = scanner.nextLong();
         
-        List<Integer> numbers = new ArrayList<>();
+        List<Long> numbers = new ArrayList<>();
         
         if(a > (b + 1))
         {
-            for(int i = b + 1; i < a; i++)
+            for(long i = b + 1; i < a; i++)
             {
                 numbers.add(i);
             }
             System.out.println((a - b - 1));
         } else if((a + 1) < b)
         {
-            for(int i = a + 1; i < b; i++)
+            for(long i = a + 1; i < b; i++)
             {
                 numbers.add(i);
             }
@@ -29,7 +29,7 @@ public class Main
             return;
         }
         Collections.sort(numbers);
-        for(int n : numbers)
+        for(long n : numbers)
         {
             System.out.print(n + " ");
